# static/swagger.yaml
swagger: "2.0"
info:
  title: "MovieApp"
  version: "1.0.3"
  description: API de l’application mobile MovieApp
  termsOfService: "http://swagger.io/terms/"
  contact:
    name: "MovieApp API Team"
    email: "contact@movieapp.fr"
schemes:
  - "http"
host: "api.movieapp.fr"
basePath: "/"
produces:
  - "application/json"
...
paths:
...
  /shows:
    get:
      tags:
        - "Shows"
      summary: Voir ses shows favoris
      parameters:
        - $ref: '#components/parameters/AccessToken'
      responses:
        200:
          description: Shows trouvés
          schema:
            $ref: "#/definitions/ShowsFound"
        401:
          description: Accès interdit ou token manquant/invalide
          schema:
            $ref: "#/definitions/Error"
        404:
          description: Aucun show trouvé
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Erreur interne
          schema:
            $ref: "#/definitions/Error"
  /shows/{show_id}:
    post:
      tags:
        - "Shows"
      summary: Ajouter un show à ses favoris
      parameters:
        - $ref: '#components/parameters/AccessToken'
        - name: show_id
          in: path
          description: ID du show
          required: true
          schema:
            type: string
        - name: show_type
          in: query
          description: Type du show
          required: true
          type: string
          enum: ['SERIE', 'MOVIE']
      responses:
        201:
          description: Show aimé
          schema:
            $ref: "#/definitions/Error"
        401:
          description: Accès interdit ou token manquant/invalide
          schema:
            $ref: "#/definitions/Error"
        404:
          description: Show déjà aimé ou inexistant
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Erreur interne
          schema:
            $ref: "#/definitions/Error"
definitions:
...
  ShowsFound:
    type: object
    properties:
      success:
        type: boolean
      message:
        type: string
      data:
        type: array
        items:
          type: object
          properties:
            show_id:
              type: integer
            show_name:
              type: string
            show_tag:
              type: integer
            show_type:
              type: string
              enum: ['SERIE', 'MOVIE']
            user_tag:
              type: string
...
  Error:
    type: object
    properties:
      success:
        type: boolean
      message:
        type: string